% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_boxplot.R
\name{.fn_compute_pval_brackets}
\alias{.fn_compute_pval_brackets}
\title{Compute p-value bracket positions with overlap detection}
\usage{
.fn_compute_pval_brackets(
  groups,
  box,
  compute.pval,
  pval.FUN,
  outline,
  at,
  horizontal,
  pval.values
)
}
\description{
Compute p-value bracket positions with overlap detection
}
\keyword{internal}
