% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_boxplot.R
\name{.fn_draw_pval_brackets}
\alias{.fn_draw_pval_brackets}
\title{Draw p-value brackets with formatted labels}
\usage{
.fn_draw_pval_brackets(pval, horizontal, pval.cex, pval.stars, pval.values)
}
\description{
Draw p-value brackets with formatted labels
}
\keyword{internal}
