% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_boxplot.R
\name{fn_boxplot}
\alias{fn_boxplot}
\alias{fn_boxplot.default}
\alias{fn_boxplot.matrix}
\alias{fn_boxplot.formula}
\title{Enhanced Boxplot}
\usage{
fn_boxplot(x, ...)

\method{fn_boxplot}{default}(
  x,
  ...,
  compute.pval = NULL,
  pval.FUN = function(x, y) wilcox.test(x, y)$p.value,
  pval.cex = 0.7,
  pval.stars = TRUE,
  pval.values = FALSE,
  tilt.names = FALSE,
  srt = 45,
  range = 1.5,
  width = NULL,
  varwidth = FALSE,
  notch = FALSE,
  outline = FALSE,
  pts.col = adjustcolor("lightgrey", 0.4),
  pts.cex = 0.6,
  names,
  plot = TRUE,
  border = par("fg"),
  col = NULL,
  log = "",
  pars = list(boxwex = ifelse(violin, 0.2, 0.4), staplewex = NA, outwex = NA, outpch =
    NA),
  horizontal = FALSE,
  add = FALSE,
  at = NULL,
  frame = FALSE,
  whisklty = ifelse(violin, 2, 1),
  lwd = par("lwd"),
  ylim = NULL,
  xaxt = "s",
  violin = FALSE,
  viocol = "#FFFFFF00",
  viowex = 0.4
)

\method{fn_boxplot}{matrix}(x, use.cols = TRUE, ...)

\method{fn_boxplot}{formula}(formula, data = NULL, ..., subset, na.action = NULL)
}
\arguments{
\item{x}{Formula, list, matrix, or data passed to boxplot.}

\item{...}{Extra parameters passed to boxplot().}

\item{compute.pval}{List of length-2 vectors with group indices to compare.
E.g., list(c(1,2), c(2,3)) compares groups 1v2 and 2v3.}

\item{pval.FUN}{Function to compute p-value. Takes two numeric vectors, returns p-value.
Default: function(x, y) wilcox.test(x, y)$p.value.}

\item{pval.cex}{Cex for p-value text. Default: 0.7.}

\item{pval.stars}{Show significance stars? Default: TRUE.}

\item{pval.values}{Show numeric p-values? Default: FALSE.}

\item{tilt.names}{Tilt x-axis labels? Default: FALSE.}

\item{srt}{Rotation angle for tilted names. Default: 45.}

\item{outline}{Show outliers? Default: FALSE.}

\item{pts.col}{Outlier point color. Default: adjustcolor("lightgrey", 0.4).}

\item{pts.cex}{Outlier point size. Default: 0.6.}

\item{violin}{Show violin density polygons behind boxes? Default: FALSE.}

\item{viocol}{Violin fill color. Default: transparent.}

\item{viowex}{Violin width expansion factor. Default: 0.4.}
}
\description{
Wrapper around base boxplot with cleaner defaults, optional violin overlay,
jittered outlier points, and p-value brackets between groups.
}
\section{Methods (by class)}{
\itemize{
\item \code{fn_boxplot(default)}: Default method for lists and vectors

\item \code{fn_boxplot(matrix)}: Method for matrices

\item \code{fn_boxplot(formula)}: Method for formulas

}}
\examples{
# Basic usage with p-values
fn_boxplot(len ~ dose, data = ToothGrowth,
           col = c("#e76f51", "#2a9d8f", "#264653"),
           compute.pval = list(c(1,2), c(2,3), c(1,3)))

# Violin mode
fn_boxplot(len ~ dose, data = ToothGrowth,
           violin = TRUE, viocol = adjustcolor("#75AAD8", 0.3),
           col = "#75AAD8")

}
